Moved note- and date handling (skip) to delete-threads

+++ script.py
@@ -50,6 +50,13 @@ def quick_add_task(task_string):
 def delete_task(task):
     task.delete()
 
+def delete_task_if_no_notes(task):
+    if len(task.get_notes()) != 0: # No note_handling implemented, skipping
+        l.info("{} has note(s), skipping".format(task.content))
+        spawn_process(task_to_project, (task, inbox_tasks))
+    else:
+        task.delete()
+
 def update_and_task_to_project(task, project):
     task.update()
     task.move(project)
@@ -71,6 +78,11 @@ def spawn_process(function, args):
 def add_improvement(task):
     airtable.insert({"Description": task.content,
                      "Phase": "Incubating"})
+    if len(task.get_notes()) != 0: # No note_handling implemented, skipping
+        l.info("{} has note(s), skipping".format(task.content))
+        spawn_process(task_to_project, (task, inbox_tasks))
+        return
+
     task.delete()
 
 def process_no_prefix(task):
@@ -117,8 +129,9 @@ def process_no_prefix(task):
         task_string += " #Inbox-tasks"
 
     l.info("Adding: \n'{}'".format(task_string))
+
     spawn_process(quick_add_task, (task_string,))
-    spawn_process(delete_task, (task,))
+    spawn_process(delete_task_if_no_notes, (task,))
 
 def process_prefixed(task):
     l.info("Processing as prefixed")
@@ -154,16 +167,11 @@ i = 0
 
 for task in due_tasks:
     i += 1
-    if len(task.get_notes()) != 0:
-        l.info("{} has note(s), skipping".format(task.content))
-        spawn_process(task_to_project, (task, inbox_tasks))
-        continue
 
     if task.due_date_utc is not None: # No due_date handling implemented, if task has due date, move and skip
         l.info("{} has a due date, skipping".format(task.content))
         spawn_process(task_to_project, (task, inbox_tasks))
         continue
-
     elif task.content[0:3] == "F: ": # If flashcard, just send straight to flashcards. Pruning can happen later.
         l.info("{} is a flashcard, moving".format(task.content))
         spawn_process(process_prefixed, (task,))
